Removed Login component;
